requirements.txt
Telethon==1.10.10
python-dotenv==v0.8.2